refactor, song fragment
