+++ app/src/main/java/uz/suxa/metaworship/presentation/SongFragment.kt
@@ -1,0 +1,32 @@
+package uz.suxa.metaworship.presentation
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import uz.suxa.metaworship.databinding.FragmentSongBinding
+import uz.suxa.metaworship.presentation.viewmodel.SongViewModel
+
+class SongFragment : Fragment() {
+
+    private var _binding: FragmentSongBinding? = null
+    private val binding get() = _binding!!
+
+    private val viewModel by lazy {
+        ViewModelProvider(
+            this,
+            ViewModelProvider.AndroidViewModelFactory.getInstance(requireActivity().application)
+        )[SongViewModel::class.java]
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentSongBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+}
